add TDM_TO_OSM
to export street emissions to Munich

TDM streets represent a macro simulation, that is,  streets sometimes are a simplification of real streets, or have less streets than the reality. Therefore, emissions can be passed to open street map like this:

- EMISSIONS_TDM_STREETS -> EMISSIONS_GRID
- EMISSIONS_GRID -> EMISSIONS_OSM_STREET
Then create .rds and .csv for convenience but the author is free to change anything

This script does that with the packages vein and eixport. This scirpt also explodes the lines in each vertex